requirements.txt
discord.py
groq
python-dotenv
streamlit
supabase